--- Pipfile
@@ -1,12 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[dev-packages]
-black = "*"
-mypy = "*"
-pytest = "*"
-
-[pipenv]
-allow_prereleases = true
